client/requirements.txt
# PointStream Client Requirements
# Additional dependencies for the client-side generative models

# Core PyTorch extensions for advanced models
torchsummary
torchmetrics
lpips  # For perceptual loss

# Video processing
moviepy
imageio
imageio-ffmpeg

# Quality assessment
scikit-image  # For SSIM/PSNR

# Optional experiment tracking
wandb  # For experiment tracking
tensorboard  # For monitoring training

# Development tools
python-dotenv
ipykernel  # Jupyter kernel support
